Fixes for the ActiveRecordTranslator
* Move `first` and `all` at the end of queries
* Satisfy postgres need that ORDER BY columns be part of the SELECT list
* Handle complex ANDs/ORs with many NOTs

There are still some issues in ActiveRecord queries.
Most Importantly, we need well typed-queries for all backends because,
in most cases, the computations are made in ruby. For example,
consider the query
`Snappy.where(["snappy.chernobyl >= ?", ('Clint'/'Julius')]).all`,
the division between the two strings in ruby will produce an error.
Another issue, is the integer division for mysql queries.
For instance, the query
`Droopier.where(["droopier.deputizes >= ?", (3/2)])`
will produce different results from queries produced by different
orms because the 3/2 division will be computed in ruby part and not
in mysql.

We tested the ActiveRecordTranslator extensively,
thus this commit closes #24.